No idea if any of this is effective, but...
Wrote sendMessageToUser, sendMessageToConversation, editMessage, and deleteMessage methods in User.

Additionally added some small additions like getName in User, private static ArrayList of Conversation and getConversationList and addConversation method in AccountHandler.

I kind of made assumptions on some of this, so if there is anything in need of edit, let me know ASAP.